patentable-project/requirements.txt
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
xgboost==1.7.6
torch==2.0.1
transformers==4.31.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4
PyMuPDF==1.23.1
pdfplumber==0.9.0
nltk==3.8.1
regex==2023.6.3


# FastAPI Backend Requirements
# Core FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database and ORM
sqlalchemy==2.0.23
alembic==1.12.1

# Data processing and ML
pandas==2.1.3
numpy==1.24.4
scikit-learn==1.3.2
sentence-transformers==2.2.2
transformers==4.35.2
torch==2.1.1

# Vector database and search
faiss-cpu==1.7.4
chromadb==0.4.15

# HTTP client and async support
httpx==0.25.2
aiofiles==23.2.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
requests==2.31.0

# Development and code quality
black==23.10.1
isort==5.12.0
flake8==6.1.0
mypy==1.7.0

# Logging and monitoring
loguru==0.7.2
prometheus-client==0.19.0

# Configuration and environment
python-dotenv==1.0.0
pydantic-settings==2.1.0

# Security
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

# Additional utilities
python-dateutil==2.8.2
pytz==2023.3